move review states type definition to /templates/feedback
