+++ apps/src/templates/instructions/teacherFeedback/types.js
@@ -1,5 +1,4 @@
 import PropTypes from 'prop-types';
-import {makeEnum} from '@cdo/apps/utils';
 
 export const teacherFeedbackShape = PropTypes.shape({
   comment: PropTypes.string,
@@ -16,9 +15,3 @@ export const rubricShape = PropTypes.shape({
   performanceLevel3: PropTypes.string,
   performanceLevel4: PropTypes.string
 });
-
-export const ReviewStates = makeEnum(
-  'completed',
-  'keepWorking',
-  'awaitingReview'
-);
